// src/types/navigation.ts
export type Screen =
  | 'splash'
  | 'login'
  | 'signup'
  | 'home'
  | 'chat'
  | 'profile'
  | 'settings'
  | 'settings_language'
  | 'settings_text_size'
  | 'settings_manage_data'
  | 'settings_privacy_policy'
  | 'forgot_password'
  | 'upgrade'
  | 'history'
  | 'edit_profile'
  | 'help'
  | 'about'
  | 'disclaimer';